requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
aiofiles==24.1.0
python-dotenv==1.0.0
sqlite-utils==3.35.2
httpx==0.25.2
websockets==12.0